+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/before/HomeBeforeChallenge.kt
@@ -1,19 +1,13 @@
-package com.mashup.twotoo.presenter.home
+package com.mashup.twotoo.presenter.home.before
 
-import androidx.annotation.DrawableRes
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.times
@@ -23,7 +17,7 @@ import com.mashup.twotoo.presenter.designsystem.component.TwoTooImageView
 import com.mashup.twotoo.presenter.designsystem.component.button.TwoTooTextButton
 import com.mashup.twotoo.presenter.designsystem.theme.TwoTooTheme
 import com.mashup.twotoo.presenter.home.model.BeforeChallengeUiModel
-import com.mashup.twotoo.presenter.home.model.StateTitleUiModel
+import com.mashup.twotoo.presenter.home.ongoing.HomeGoalCount
 
 /**
  * @Created by 김현국 2023/06/10
@@ -83,45 +77,6 @@ fun HomeBeforeChallenge(
             text = stringResource(id = beforeChallengeUiModel.buttonText),
         )
     }
-}
-
-@Composable
-fun HomeBeforeChallengeTitle(
-    stateTitle: StateTitleUiModel,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(28.dp),
-    ) {
-        Text(
-            text = stringResource(id = stateTitle.title),
-            style = TwoTooTheme.typography.headLineNormal28,
-            color = TwoTooTheme.color.mainBrown,
-            textAlign = TextAlign.Center,
-        )
-        stateTitle.subTitle?.let {
-            Text(
-                text = stringResource(id = stateTitle.subTitle),
-                style = TwoTooTheme.typography.bodyNormal14,
-                color = TwoTooTheme.color.gray600,
-                textAlign = TextAlign.Center,
-            )
-        }
-    }
-}
-
-@Composable
-fun HomeBeforeChallengeImage(
-    @DrawableRes stateImage: Int,
-    modifier: Modifier = Modifier,
-) {
-    TwoTooImageView(
-        modifier = modifier,
-        model = stateImage,
-        previewPlaceholder = stateImage,
-    )
 }
 
 @Preview(showBackground = true, device = "id:pixel_2")
