+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/ongoing/HomeShotCountText.kt
@@ -1,4 +1,4 @@
-package com.mashup.twotoo.presenter.home
+package com.mashup.twotoo.presenter.home.ongoing
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
